Created js and jsx files of images for the map marker and separated pin and ball icons so that colors can be changed using inline CSS. - CourtFinderMarker, BallIcon.jsx, and PinIcon.jsx.
Marker was added in to Map.jsx.
Created inputs for busyness and courtType to see how it would look on the map.

Used Legend.css to create a legend with orange borders and detailed icons/descriptions.
Also restructured Legend.jsx to incorporate the change to using js and jsx files instead of the svgs themselves.
Added class names and used courtStatus.js to change color for each icon of the legend.

The image for the ball icon svg was also changed because the old one didnt have paths that could be edited.
The main marker image was changed to use the new ball icon as well.
Removed old court finder pin.

Renamed the old svg's that were being used for images - they're no longer being used
but files are kept in /assets for reference.